// src/Controllers/ConversationalCloud/conversation-cloud.interfaces.ts
export interface ConversationSurvey {
  surveyType: string;
  surveyStatus: string;
  dialogId: string;
  surveyData: {
    answer: string;
    answerId: string;
    answerSeq: number;
    isValidAnswer: boolean;
    question: string;
    questionFormat: string;
    questionId: string;
    questionName: string;
    questionType: string;
  }[];
}
interface CoBrowseSession {
  sessionId: string;
  startTime: string;
  startTimeL: number;
  endTime: string;
  endTimeL: number;
  endReason: string;
  duration: number;
  type: string;
  capabilities?: string[];
  agentId: string;
  interactive: boolean;
}
interface CoBrowseSessions {
  coBrowseSessionsList: CoBrowseSession[];
}

export interface CustomerInfoSde {
  serverTimeStamp: string;
  customerInfo: {
    companyBranch: string;
    customerId: string;
    customerStatus: string;
    customerType: string;
  };
}

export interface SDE {
  customerInfo?: CustomerInfoSde;
  personalInfo?: {
    originalTimeStamp: string;
    personalInfo: {
      contacts: {
        personalContact: {
          email: string;
          phone: string;
        };
      }[];
      name: string;
      surname: string;
    };
    serverTimeStamp: string;
  };
  sdeType: string;
  serverTimeStamp: string;
}

export interface SDEs {
  events: SDE[];
}